Extract CoreData managed classes into separate files
