+++ EssentialFeed/EssentialFeed/Feed Cache/Infrastructure/CoreData/CoreDataFeedStore.swift
@@ -73,54 +73,3 @@ public final class CoreDataFeedStore: FeedStore {
         }
     }
 }
-
-@objc(ManagedCache)
-private class ManagedCache: NSManagedObject {
-    @NSManaged var timestamp: Date
-    @NSManaged var feed: NSOrderedSet
-    
-    var localFeed: [LocalFeedImage] {
-        return feed.compactMap { ($0 as? ManagedFeedImage)?.local }
-    }
-    
-    static func find(in context: NSManagedObjectContext) throws -> ManagedCache? {
-        let request = NSFetchRequest<ManagedCache>(entityName: entity().name!)
-        request.returnsObjectsAsFaults = false
-        return try context.fetch(request).first
-    }
-    
-    static func newUniqueInstance(in context: NSManagedObjectContext) throws -> ManagedCache {
-        try find(in: context).map(context.delete)
-        return ManagedCache(context: context)
-    }
-}
-
-@objc(ManagedFeedImage)
-private class ManagedFeedImage: NSManagedObject {
-    @NSManaged var id: UUID
-    @NSManaged var imageDescription: String?
-    @NSManaged var location: String?
-    @NSManaged var url: URL
-    @NSManaged var cache: ManagedCache
-    
-    static func images(from localFeed: [LocalFeedImage], in context: NSManagedObjectContext) -> NSOrderedSet {
-        let managedFeedImages = localFeed.map {
-            let managedFeedImage = ManagedFeedImage(context: context)
-            managedFeedImage.id = $0.id
-            managedFeedImage.imageDescription = $0.description
-            managedFeedImage.location = $0.location
-            managedFeedImage.url = $0.url
-            return managedFeedImage
-        }
-        return NSOrderedSet(array: managedFeedImages)
-    }
-    
-    var local: LocalFeedImage {
-        LocalFeedImage(
-            id: id,
-            description: imageDescription,
-            location: location,
-            url: url
-        )
-    }
-}
