+++ EssentialFeed/EssentialFeed/Feed Cache/Infrastructure/CoreData/ManagedFeedImage.swift
@@ -1,0 +1,38 @@
+//
+//  ManagedFeedImage.swift
+//  EssentialFeed
+//
+//  Created by Cristian Felipe Patiño Rojas on 9/4/25.
+//
+
+import CoreData
+
+@objc(ManagedFeedImage)
+class ManagedFeedImage: NSManagedObject {
+    @NSManaged var id: UUID
+    @NSManaged var imageDescription: String?
+    @NSManaged var location: String?
+    @NSManaged var url: URL
+    @NSManaged var cache: ManagedCache
+    
+    static func images(from localFeed: [LocalFeedImage], in context: NSManagedObjectContext) -> NSOrderedSet {
+        let managedFeedImages = localFeed.map {
+            let managedFeedImage = ManagedFeedImage(context: context)
+            managedFeedImage.id = $0.id
+            managedFeedImage.imageDescription = $0.description
+            managedFeedImage.location = $0.location
+            managedFeedImage.url = $0.url
+            return managedFeedImage
+        }
+        return NSOrderedSet(array: managedFeedImages)
+    }
+    
+    var local: LocalFeedImage {
+        LocalFeedImage(
+            id: id,
+            description: imageDescription,
+            location: location,
+            url: url
+        )
+    }
+}
